Changes to mempool interface based on PR feedback

+++ libdialog/dialog/configuration_params.h
@@ -29,7 +29,7 @@ static utils::configuration_map dialog_conf(
     utils::config_utils::read_from_env(
         "DIALOG_CONF", "/etc/conf/dialog.conf:./conf/dialog.conf"));
 
-size_t configuration_params::MAX_MEMORY = dialog_conf.get<int>(
+size_t configuration_params::MAX_MEMORY = dialog_conf.get<size_t>(
     "max_memory", constants::DEFAULT_MAX_MEMORY);
 int configuration_params::MAX_CONCURRENCY = dialog_conf.get<int>(
     "max_concurrency", constants::HARDWARE_CONCURRENCY);
